requirements/base.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.15
django-cors-headers==2.4.0
django-environ==0.4.5
django-extensions==2.1.2
django-filter==1.1.0
django-grappelli==2.11.1
django-rest-swagger==2.1.2
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
drf-openapi==1.3.0
drf-yasg==1.9.1
flex==6.13.2
functools32==3.2.3.post2
future==0.16.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.7.5
pytz==2018.5
PyYAML==3.13
requests==2.19.1
rfc3987==1.3.8
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.66
simplejson==3.16.1
six==1.11.0
sqlparse==0.2.4
strict-rfc3339==0.7
swagger-spec-validator==2.3.1
typing==3.6.6
uritemplate==3.0.0
urllib3==1.23
validate-email==1.3
whitenoise==4.0
